services/payment-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
aiokafka==0.11.0
python-dotenv==1.0.0
pydantic-settings==2.1.0
httpx==0.25.1
pytest==7.4.3
pytest-asyncio==0.21.1
respx==0.20.2
aiosqlite==0.19.0